HIR-44 refactor: 북마크 서비스 계층도 Company와 Job으로 각각 분리함
